+++ jest.config.js
@@ -10,7 +10,6 @@ export default {
   ],
   projects: [
     '<rootDir>/packages/i18n',
-    '<rootDir>/packages/oembed',
     '<rootDir>/packages/portal',
     '<rootDir>/packages/vue-router-query',
     '<rootDir>/packages/vue-session',
