+++ package.json
@@ -35,7 +35,6 @@
   "resolutions": {
     "@europeana/contentful": "link:./packages/contentful",
     "@europeana/i18n": "link:./packages/i18n",
-    "@europeana/oembed": "link:./packages/oembed",
     "@europeana/portal": "link:./packages/portal",
     "@europeana/style": "link:./packages/style",
     "@europeana/vue-router-query": "link:./packages/vue-router-query",
