--- packages/oembed/babel.config.cjs
@@ -1,9 +1,0 @@
-// Babel config used by Jest, e.g. for unit tests
-module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      { targets: { node: 'current' } }
-    ]
-  ]
-};
