--- packages/oembed/jest.config.js
@@ -1,18 +1,0 @@
-export default {
-  displayName: '@europeana/oembed',
-  moduleFileExtensions: [
-    'js',
-    'json',
-    'mjs'
-  ],
-  moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/src/$1'
-  },
-  testEnvironment: 'jsdom',
-  testPathIgnorePatterns: [
-    '/node_modules/'
-  ],
-  transform: {
-    '^.+\\.(js|mjs)$': 'babel-jest'
-  }
-};
