--- packages/oembed/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@europeana/oembed",
-  "version": "2.0.0",
-  "description": "Europeana oEmbed client & providers",
-  "license": "EUPL-1.2",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "main": "./src/index.js",
-  "dependencies": {
-    "axios": "^0.21.4",
-    "lodash": "^4.17.21"
-  }
-}
