+++ packages/portal/package.json
@@ -16,7 +16,6 @@
   "dependencies": {
     "@elastic/apm-rum-vue": "^1.3.1",
     "@europeana/i18n": "^2.0.0",
-    "@europeana/oembed": "^2.0.0",
     "@europeana/vue-router-query": "^2.0.0",
     "@europeana/vue-session": "^2.0.0",
     "@europeana/vue-visible-on-scroll": "^2.0.0",
